refactor: centralize `Encoder` class in encoder.py file
