+++ letterboxdpy/encoder.py
@@ -1,0 +1,10 @@
+from json import JSONEncoder
+
+class Encoder(JSONEncoder):
+    """
+    Encoder class provides a way to serialize custom class
+    .. instances to JSON by overriding the default serialization
+    .. logic to return the object's namespace dictionary.
+    """
+    def default(self, o):
+        return o.__dict__
