+++ letterboxdpy/list.py
@@ -1,9 +1,9 @@
 from letterboxdpy.scraper import Scraper
+from letterboxdpy.encoder import Encoder
 from functools import wraps
 import re
 
 from json import (
-  JSONEncoder,
   dumps as json_dumps,
   loads as json_loads
 )
@@ -93,15 +93,6 @@ class List:
             page += 1
 
         return movie_list
-
-class Encoder(JSONEncoder):
-    """
-    Encoder class provides a way to serialize custom class
-    .. instances to JSON by overriding the default serialization
-    .. logic to return the object's namespace dictionary.
-    """
-    def default(self, o):
-        return o.__dict__
 
 # -- DECORATORS --
